Add logic of creating groups

+++ src/main/java/com/example/easyDocs/Document/DocumentService.java
@@ -1,7 +1,9 @@
 package com.example.easyDocs.Document;
 
+import com.example.easyDocs.AccessGroup.AccessGroupService;
 import com.example.easyDocs.User.User;
 import com.example.easyDocs.User.UserRepository;
+import com.example.easyDocs.exceptions.AccessException;
 import com.example.easyDocs.exceptions.DocumentException;
 import com.example.easyDocs.exceptions.UserNotFoundException;
 import org.springframework.core.io.Resource;
@@ -11,7 +13,9 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
 import java.time.LocalDate;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 @Component
@@ -22,12 +26,22 @@ public class DocumentService {
     DocumentStorageService documentStorageService;
     UserRepository userRepository;
     List<String> supported_files = List.of("txt", "doc", "docx", "pdf");
+    AccessGroupService accessGroupService;
 
-    public DocumentService(DocumentMapper documentMapper, DocumentRepository documentRepository, DocumentStorageService documentStorageService, UserRepository userRepository){
+    public DocumentService(DocumentMapper documentMapper, DocumentRepository documentRepository,
+                           DocumentStorageService documentStorageService, UserRepository userRepository,
+                           AccessGroupService accessGroupService){
+
         this.documentMapper = documentMapper;
         this.documentRepository = documentRepository;
         this.documentStorageService = documentStorageService;
         this.userRepository = userRepository;
+        this.accessGroupService = accessGroupService;
+    }
+
+    public User getAuthenticatedUser(Authentication authentication){
+        String email = authentication.getName();
+        return userRepository.findByEmail(email).orElseThrow(() -> new UserNotFoundException(email));
     }
 
     public List<DocumentDto> getDocuments() {
@@ -36,33 +50,74 @@ public class DocumentService {
                 .collect(Collectors.toList());
     }
 
-    public DocumentDto getDocumentById(Long id){
-        Document file = documentRepository.findById(id).orElse(new Document());
-        return documentMapper.documentToDocumentDto(file);
+    public DocumentDto getDocumentById(Long id, Authentication authentication){
+        Document document = documentRepository.findById(id).orElseThrow(() -> new DocumentException(id));
+
+        User user = getAuthenticatedUser(authentication);
+
+        if(accessGroupService.haveAccess(document, user) || user.getRole().equals("ADMIN")){
+            return documentMapper.documentToDocumentDto(document);
+        } else {
+            throw new AccessException();
+        }
     }
 
-    public List<DocumentDto> getDocumentsByName(String name){
-        List<Document> documents = documentRepository.findAllByName(name);
-        return documents
+    public List<DocumentDto> getDocumentsByName(String name,Authentication authentication){
+
+        User user = getAuthenticatedUser(authentication);
+
+        Set<Document> documents = documentRepository.findAllByName(name);
+        Set<Document> haveAccessDocuments = new HashSet<>();
+
+        for(Document document: documents){
+            if(accessGroupService.haveAccess(document, user) || user.getRole().equals("ADMIN")){
+                haveAccessDocuments.add(document);
+            }
+        }
+
+        return haveAccessDocuments
                 .stream()
                 .map(document -> documentMapper.documentToDocumentDto(document))
                 .collect(Collectors.toList());
     }
 
-    public void deleteDocument(Long id) {
+    public void deleteDocument(Long id, Authentication authentication) {
+
         Document document = documentRepository.findById(id).orElseThrow(() -> new DocumentException(id));
-        documentRepository.delete(document);
+        User user = getAuthenticatedUser(authentication);
+
+        if(document.getCreator().equals(user) || user.getRole().equals("ADMIN")) {
+            documentRepository.delete(document);
+        } else {
+            throw new AccessException();
+        }
     }
 
-    public Resource getDocumentAsResource(Long id){
+    public Resource getDocumentAsResource(Long id, Authentication authentication){
         Document document = documentRepository.findById(id).orElseThrow(() -> new DocumentException(id));
-        return documentStorageService.getDocumentAsResource(document.getFile_path());
+        User user = getAuthenticatedUser(authentication);
+
+        if(accessGroupService.haveAccess(document, user) || user.getRole().equals("ADMIN")) {
+            return documentStorageService.getDocumentAsResource(document.getFile_path());
+        } else {
+            throw new AccessException();
+        }
     }
 
-    public List<DocumentDto> getDocumentsByCreator(Long creator_id){
-        User user = userRepository.findById(creator_id).orElseThrow(() -> new UserNotFoundException(creator_id));
-        List<Document> documents = documentRepository.findAllByCreator(user);
-        return documents.stream()
+    public List<DocumentDto> getDocumentsByCreator(Long creator_id, Authentication authentication){
+        User user = getAuthenticatedUser(authentication);
+        User creator = userRepository.findById(creator_id).orElseThrow(() -> new UserNotFoundException(creator_id));
+
+        Set<Document> documents = documentRepository.findAllByCreator(creator);
+        Set<Document> haveAccessDocuments = new HashSet<>();
+
+        for(Document document: documents){
+            if(accessGroupService.haveAccess(document, user) || user.getRole().equals("ADMIN")){
+                haveAccessDocuments.add(document);
+            }
+        }
+
+        return haveAccessDocuments.stream()
                 .map(document -> documentMapper.documentToDocumentDto(document))
                 .collect(Collectors.toList());
     }
@@ -98,19 +153,26 @@ public class DocumentService {
         documentRepository.save(document);
     }
 
-    public void updateDocument(Long id, Document document) {
-        Document documentUpdate = documentRepository.findById(id).orElseThrow(() -> new DocumentException(id));
+    public void updateDocument(Long id, Document document, Authentication authentication) {
+
+        User user = getAuthenticatedUser(authentication);
+        Document documentToUpdate = documentRepository.findById(id).orElseThrow(() -> new DocumentException(id));
+
+        if (!documentToUpdate.getCreator().equals(user)){
+            throw new AccessException();
+        }
+
         boolean documentChanged = false;
 
         if(document.getName()!=null){
-            documentUpdate.setName(document.getName());
+            documentToUpdate.setName(document.getName());
             documentChanged = true;
         }
         if(document.getDescription()!=null){
-            documentUpdate.setDescription(document.getDescription());
+            documentToUpdate.setDescription(document.getDescription());
             documentChanged = true;
         }
-        if(documentChanged){documentUpdate.setLastUpdate(LocalDate.now());}
-        documentRepository.save(documentUpdate);
+        if(documentChanged){documentToUpdate.setLastUpdate(LocalDate.now());}
+        documentRepository.save(documentToUpdate);
     }
 }
